RQ2-Topics_of_blogs/LDA_python_scripts/requirements.txt
beautifulsoup4==4.11.1
matplotlib==3.5.3
nltk==3.7
openpyxl==3.0.10
pandas==1.4.3
pyenchant==3.2.2
requests==2.28.1
sklearn==0.0
tqdm==4.64.0
